requirements.txt
git+https://github.com/huggingface/transformers
datasets
sentencepiece
accelerate
evaluate
python-Levenshtein
rouge-score